activity files and folders naming improved. I made them all in lowercase
