+++ app/src/main/java/com/example/test/activity/records/MyTimeRecordsActivity.kt
@@ -1,4 +1,4 @@
-package com.example.test.activity.myRecords
+package com.example.test.activity.records
 
 import android.content.Context
 import android.content.Intent
@@ -9,41 +9,42 @@ import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.RecyclerView
+import com.example.test.adapter.TimeRecordItemAdapter
 import com.example.test.R
 import com.example.test.activity.basics.HomeActivity
 import com.example.test.activity.basics.MyPersonalProfileActivity
-import com.example.test.activity.interactionOfUsers.FindUserActivity
-import com.example.test.adapter.SetsRecordItemAdapter
+import com.example.test.activity.interactions.FindUserActivity
 import com.example.test.data.Datasource
+import com.example.test.model.TimesRecord
 import kotlinx.coroutines.launch
 
-class MySetsRecordsActivity : AppCompatActivity() {
+class MyTimeRecordsActivity : AppCompatActivity() {
 
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_my_sets_records)
-        val addBtn: Button =findViewById(R.id.my_sets_records_AddRecordBtn)
-        val homeBtn: Button =findViewById(R.id.my_sets_records_HomeBtn)
+        setContentView(R.layout.activity_my_time_records)
 
+        val homeBtn: Button = findViewById(R.id.my_time_records_HomeBtn)
+        val addBtn: Button = findViewById(R.id.my_time_records_AddRecordBtn)
         val sharedPreferences = getSharedPreferences("account_info", Context.MODE_PRIVATE)
         val account = sharedPreferences.getString("account", "")
         val context = this
-        val recyclerView = findViewById<RecyclerView>(R.id.my_sets_records_RecyclerView)
+        val recyclerView = findViewById<RecyclerView>(R.id.my_time_records_RecycleView)
 
         addBtn.setOnClickListener {
-            val intent = Intent(context, AddSetsRecordsActivity::class.java)
+            val intent = Intent(context, AddTimeRecordsActivity::class.java)
             context.startActivity(intent)
         }
+
         homeBtn.setOnClickListener {
             val intent = Intent(context, HomeActivity::class.java)
             context.startActivity(intent)
         }
 
         lifecycleScope.launch {
-            val data = Datasource().loadSetsRecords(account.toString())
-
-            recyclerView.adapter = SetsRecordItemAdapter(data)
+            val data: List<TimesRecord> = Datasource().loadTimesRecords(account.toString())
+            recyclerView.adapter = TimeRecordItemAdapter(dataset = data)
             recyclerView.setHasFixedSize(true)
         }
 
@@ -63,6 +64,7 @@ class MySetsRecordsActivity : AppCompatActivity() {
 
                 true
             }
+
             R.id.menu_MyProfile -> {
                 val intent = Intent(context, MyPersonalProfileActivity::class.java)
                 context.startActivity(intent)
