+++ app/src/main/res/layout/activity_add_time_records.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.myRecords.AddTimeRecordsActivity">
+    tools:context=".activity.records.AddTimeRecordsActivity">
 
     <TextView
         android:id="@+id/add_time_text"
